futureagi/model_serving/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "model-serving"
version = "0.1.0"
description = "Model serving application"
license = { text = "MIT" }
authors = [
    { name = "Future AGI Team", email = "team@futureagi.com" }
]
maintainers = [
    { name = "Future AGI Team", email = "team@futureagi.com" }
]
keywords = ["machine-learning", "embeddings", "fastapi", "model-serving"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"

# Core dependencies
dependencies = [
    "fastapi==0.115.14",
    "starlette==0.46.2",
    "pydantic==1.10.18",
    "gunicorn==21.2.0",
    "uvicorn==0.23.1",
    "uvloop==0.20.0",
    "watchfiles==0.24.0",
    "torch==2.6.0",
    "transformers==4.44.2",
    "sentence-transformers==3.3.1",
    "tokenizers==0.19.1",
    "safetensors==0.4.4",
    "scikit-learn==1.5.1",
    "scipy==1.14.1",
    "numpy==1.26.4",
    "hdbscan",
    "librosa==0.10.1",
    "soundfile==0.12.1",
    "pydub==0.25.1",
    "ffmpeg-python",
    "wav2clip==0.1.0",
    "av",
    "pdfminer.six",
    "python-docx",
    "striprtf",
    "pdfminer",
    "pypdf",
    "json-repair",
    "partial-json-parser",
    "tenacity",
    "retrying",
    "watchdog",
    "jsonpath-ng",
    "nltk",
    "rouge-score",
    "python-levenshtein",
    "chevron",
    "filetype",
    "openai",
    "requests==2.32.3",
    "urllib3==2.2.2",
    "httptools==0.6.1",
    "h11==0.16.0",
    "httpx==0.28.1",
    "anyio==4.4.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "filelock==3.15.4",
    "fsspec==2024.6.1",
    "huggingface-hub==0.24.6",
    "idna==3.8",
    "Jinja2==3.1.4",
    "joblib==1.4.2",
    "MarkupSafe==2.1.5",
    "mpmath==1.3.0",
    "networkx==3.3",
    "packaging==24.1",
    "pillow==10.4.0",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.2",
    "regex==2024.7.24",
    "sniffio==1.3.1",
    "sympy==1.13.1",
    "threadpoolctl==3.5.0",
    "tqdm==4.66.5",
    "typing_extensions==4.12.2",
    "sentry_sdk",
    "structlog>=24.1.0",
    "pandas>=2.0.0",
    # OpenTelemetry - core
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-semantic-conventions",
    # OpenTelemetry - propagators (for context propagation between services)
    "opentelemetry-propagator-b3>=1.20.0",
    # OpenTelemetry - exporters
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-exporter-otlp-proto-http",
    # OpenTelemetry - instrumentation
    "opentelemetry-instrumentation>=0.41b0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    # OpenTelemetry instrumentation - asyncio (for async context propagation)
    "opentelemetry-instrumentation-asyncio>=0.41b0",
]

# Optional dependencies
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.28.1",  # For testing FastAPI endpoints
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.971",
    "pre-commit>=2.20.0",
]

monitoring = [
    "newrelic",
    "prometheus-client",
    # OpenTelemetry packages are now in main dependencies
]

gpu = [
    "torch[cuda]",
    "torchvision",
    "torchaudio",
    "xformers",
]

# Project URLs
[project.urls]
Homepage = "https://github.com/futureagi/model-serving"
Documentation = "https://model-serving.readthedocs.io"
Repository = "https://github.com/futureagi/model-serving.git"
"Bug Tracker" = "https://github.com/futureagi/model-serving/issues"

# Entry points for CLI commands
[project.scripts]
model-serving = "app.main:main"
serve = "app.cli:serve"

# Hatch configuration
[tool.hatch.version]
path = "app/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# Development environment
[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.28.1",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.971",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov=app --cov-report=term-missing --cov-report=html {args:tests}"
lint = "flake8 app tests"
format = "black app tests && isort app tests"
type-check = "mypy app"
serve = "uvicorn app.main:app --host 0.0.0.0 --port 8080 --reload"
serve-prod = "gunicorn app.main:app -w 4 -k uvicorn.workers.UvicornWorker --bind 0.0.0.0:8080"

# Testing environment
[tool.hatch.envs.test]
dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.28.1",
]

[[tool.hatch.envs.test.matrix]]
python = ["3.10", "3.11", "3.12"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "performance: Performance tests",
    "slow: Slow running tests",
]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
show_missing = true
precision = 2

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "transformers.*",
    "sentence_transformers.*",
    "librosa.*",
    "soundfile.*",
    "pydub.*",
    "wav2clip.*",
    "pdfminer.*",
    "docx.*",
    "striprtf.*",
    "pypdf.*",
    "hdbscan.*",
    "xformers.*",
]
ignore_missing_imports = true
